Fix the order of OS_Parameter data
up to now the OS_Parameter dictionary was being created, but there was a
mismatch in between the labels used and the values stores. Mainly
because the way the data was rearrenged relied on a "set" (which is an
unordered list) and a dictionary (which is also unordered). Now almost
the same code is used, but two lists are created, ensuring that the
labels will be assigned to the correct values. This fix makes the
python_float_fix obsolete.